+++ api/src/sessionPhase1/import/sessionPhase1ImportValidator.ts
@@ -1,8 +1,8 @@
-export function sessionPhase1ImportValidator(headers: string[]): boolean {
+export function sessionPhase1ImportValidator(headers: string[]) {
   const requiredHeaders = ["Code du centre pour la session", "Session formule", "Désignation du centre", "Effectif d'individuels", "Effectif d'élèves (CLE)"];
   const missingHeaders = requiredHeaders.filter((header) => !headers.includes(header));
+
   if (missingHeaders.length > 0) {
-    return false;
+    throw new Error(missingHeaders.join(", "));
   }
-  return true;
 }
